Address PR comments from das7pad
