+++ src/main/frontend/pipeline-console-view/pipeline-console/main/PipelineConsole.tsx
@@ -113,6 +113,7 @@ export default function PipelineConsole() {
                 selectedStage={openStage || undefined}
                 stageViewPosition={stageViewPosition}
                 onStageSelect={handleStageSelect}
+                currentRunPath={currentRunPath}
               />
             ))}
