+++ src/main/frontend/pipeline-graph-view/pipeline-graph/main/NestedPipelineGraphLayout.ts
@@ -151,89 +151,6 @@ function maxGraphNodeProp(
   return Math.max(node[prop], ...node.children.map((c) => c[prop]));
 }
 
-const STATE_PRIORITY: Record<string, number> = {
-  [Result.failure]: 0,
-  [Result.unstable]: 1,
-  [Result.aborted]: 2,
-  [Result.paused]: 3,
-  [Result.running]: 4,
-  [Result.queued]: 5,
-  [Result.not_built]: 6,
-  [Result.skipped]: 7,
-  [Result.success]: 8,
-  [Result.unknown]: 9,
-};
-
-function worstState(a: Result, b: Result): Result {
-  return (STATE_PRIORITY[a] ?? 9) <= (STATE_PRIORITY[b] ?? 9) ? a : b;
-}
-
-function isTransparentState(state: Result): boolean {
-  return state === Result.skipped || state === Result.not_built;
-}
-
-function aggregateChildState(stage: StageInfo): Result {
-  let all = stage.state;
-  let nonTransparent: Result | null = isTransparentState(stage.state)
-    ? null
-    : stage.state;
-  for (const child of stage.children) {
-    const childState = aggregateChildState(child);
-    all = worstState(all, childState);
-    if (!isTransparentState(childState)) {
-      nonTransparent =
-        nonTransparent == null
-          ? childState
-          : worstState(nonTransparent, childState);
-    }
-  }
-  return nonTransparent ?? all;
-}
-
-export function collapseSelectiveStages(
-  stages: StageInfo[],
-  collapsedIds: Set<number>,
-): StageInfo[] {
-  return stages.map((stage) => {
-    if (stage.children.length === 0) {
-      return stage;
-    }
-    if (collapsedIds.has(stage.id)) {
-      return {
-        ...stage,
-        children: [],
-        collapsedChildCount: countLeafStages(stage),
-        state: aggregateChildState(stage),
-      };
-    }
-    return {
-      ...stage,
-      children: collapseSelectiveStages(stage.children, collapsedIds),
-    };
-  });
-}
-
-function countLeafStages(stage: StageInfo): number {
-  if (stage.children.length === 0) {
-    return 1;
-  }
-  return stage.children.reduce((sum, child) => sum + countLeafStages(child), 0);
-}
-
-export function collectParentStageIds(stages: StageInfo[]): Set<number> {
-  const ids = new Set<number>();
-  function walk(list: StageInfo[]) {
-    for (const stage of list) {
-      if (stage.children.length > 0) {
-        ids.add(stage.id);
-        walk(stage.children);
-      }
-    }
-  }
-  walk(stages);
-  return ids;
-}
-
 function collectCollapsedStages(
   collapsedStages: StageInfo[],
   stages: StageInfo[],
