--- src/test/resources/io/jenkins/plugins/pipelinegraphview/configure-appearance-collapse.yml
@@ -1,7 +1,0 @@
-appearance:
-  consoleUrlProvider:
-    providers:
-      - "pipelineGraphView"
-  pipelineGraphView:
-    showGraphOnBuildPage: true
-    showGraphOnJobPage: true
